+++ pb/dfuse/eosio/codec/v1/merge_events_test.go
@@ -16,7 +16,11 @@ package pbcodec
 
 import (
 	"fmt"
+	"os"
 	"testing"
+
+	"github.com/dfuse-io/logging"
+	"go.uber.org/zap"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -43,7 +47,15 @@ func TestSort(t *testing.T) {
 	assert.False(t, evs[6].Irreversible)
 }
 
+func init() {
+	if os.Getenv("DEBUG") != "" {
+		logger, _ := zap.NewDevelopment()
+		logging.Override(logger)
+	}
+}
+
 func TestMergeTransactionEvents(t *testing.T) {
+
 	tests := []struct {
 		name           string
 		events         []*TransactionEvent
@@ -67,7 +79,7 @@ func TestMergeTransactionEvents(t *testing.T) {
 		{
 			name: "two additions, none irr, check canonical chain",
 			events: []*TransactionEvent{
-				{Id: "trx1", BlockId: "a", Irreversible: false, Event: NewTestAddEvent(1)},
+				{Id: "trx1", BlockId: "a", Irreversible: false, Event: NewTestAddEvent(1)}, // skip this one since it isn't IRR and is NOT in the longest chain
 				{Id: "trx1", BlockId: "b", Irreversible: false, Event: NewTestAddEvent(2)},
 			},
 			canonicalChain: func(t *testing.T, id string) bool {
@@ -86,8 +98,8 @@ func TestMergeTransactionEvents(t *testing.T) {
 				{Id: "trx1", BlockId: "c", Irreversible: true, Event: NewTestAddEvent(3)},
 
 				{Id: "trx1", BlockId: "d", Irreversible: false, Event: NewTestExecEvent(4)},
+				{Id: "trx1", BlockId: "c", Irreversible: true, Event: NewTestExecEvent(6)},
 				{Id: "trx1", BlockId: "e", Irreversible: false, Event: NewTestExecEvent(5)},
-				{Id: "trx1", BlockId: "f", Irreversible: true, Event: NewTestExecEvent(6)},
 			},
 			canonicalChain: func(t *testing.T, id string) bool {
 				t.Error("we said never call canonicalChain!")
@@ -104,28 +116,28 @@ func TestMergeTransactionEvents(t *testing.T) {
 		{
 			name: "multiple, select one of each, ignore dtrx cancels if execution irreversible",
 			events: []*TransactionEvent{
-				{Id: "trx1", BlockId: "a", Irreversible: false, Event: NewTestDtrxCreateEvent("1")},
-				{Id: "trx1", BlockId: "b", Irreversible: true, Event: NewTestDtrxCreateEvent("2")},
-				{Id: "trx1", BlockId: "c", Irreversible: false, Event: NewTestDtrxCreateEvent("3")},
+				{Id: "trx1", BlockId: "2a", Irreversible: false, Event: NewTestDtrxCreateEvent("1")},
+				{Id: "trx1", BlockId: "2b", Irreversible: true, Event: NewTestDtrxCreateEvent("2")},
+				{Id: "trx1", BlockId: "2c", Irreversible: false, Event: NewTestDtrxCreateEvent("3")},
 
-				{Id: "trx1", BlockId: "d", Irreversible: false, Event: NewTestExecEvent(4)},
-				{Id: "trx1", BlockId: "e", Irreversible: false, Event: NewTestExecEvent(5)},
-				{Id: "trx1", BlockId: "f", Irreversible: true, Event: NewTestExecEvent(6)},
+				// delay trx thus execution could be at a later block
+				{Id: "trx1", BlockId: "4a", Irreversible: false, Event: NewTestExecEvent(4)},
+				{Id: "trx1", BlockId: "4b", Irreversible: true, Event: NewTestExecEvent(5)},
+				{Id: "trx1", BlockId: "4c", Irreversible: false, Event: NewTestExecEvent(6)},
 
-				{Id: "trx1", BlockId: "call1", Irreversible: false, Event: NewTestDtrxCancelEvent("1")},
-				{Id: "trx1", BlockId: "call2", Irreversible: false, Event: NewTestDtrxCancelEvent("2")},
+				{Id: "trx1", BlockId: "5b", Irreversible: false, Event: NewTestDtrxCancelEvent("1")},
+				{Id: "trx1", BlockId: "5d", Irreversible: false, Event: NewTestDtrxCancelEvent("2")},
 			},
 			canonicalChain: func(t *testing.T, id string) bool {
-				if id == "call1" || id == "call2" {
+				if id == "5b" {
 					return true
 				}
-				t.Error("don't call canonicalChain otherwise")
-				return true
+				return false
 			},
 			expect: &TransactionLifecycle{
 				Id:                    "trx1",
 				TransactionStatus:     TransactionStatus_TRANSACTIONSTATUS_HARDFAIL, // no receipt, ignore
-				ExecutionTrace:        &TransactionTrace{Index: 6},
+				ExecutionTrace:        &TransactionTrace{Index: 5},
 				ExecutionIrreversible: true,
 				CreationIrreversible:  true,
 				CreatedBy:             &ExtDTrxOp{SourceTransactionId: "2"}},
@@ -133,8 +145,8 @@ func TestMergeTransactionEvents(t *testing.T) {
 		{
 			name: "cancellation arrives before irreversible execution, should not show cancelled at all",
 			events: []*TransactionEvent{
-				{Id: "trx1", BlockId: "d", Irreversible: false, Event: NewTestDtrxCancelEvent("1")},
-				{Id: "trx1", BlockId: "f", Irreversible: true, Event: NewTestExecEvent(6)},
+				{Id: "trx1", BlockId: "d", BlockNum: 4, Irreversible: false, Event: NewTestDtrxCancelEvent("1")},
+				{Id: "trx1", BlockId: "f", BlockNum: 5, Irreversible: true, Event: NewTestExecEvent(6)},
 			},
 			canonicalChain: func(t *testing.T, id string) bool {
 				return true
@@ -143,6 +155,175 @@ func TestMergeTransactionEvents(t *testing.T) {
 				Id:                    "trx1",
 				TransactionStatus:     TransactionStatus_TRANSACTIONSTATUS_HARDFAIL, // no receipt, ignore
 				ExecutionTrace:        &TransactionTrace{Index: 6},
+				ExecutionIrreversible: true,
+			},
+		},
+		{
+			name: "dev1: deferred transaction push, has multiple execution traces, execution succeeded",
+			events: []*TransactionEvent{
+				{
+					Id:           "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					BlockId:      "0000002ca98c204ec8f93e91332baf2a0eea5edd4e9f262bfb3e9b997d6f2415",
+					Irreversible: true,
+					Event: &TransactionEvent_Addition{Addition: &TransactionEvent_Added{
+						Receipt: &TransactionReceipt{Index: uint64(1)},
+						Transaction: &SignedTransaction{
+							Transaction: &Transaction{
+								Actions: []*Action{
+									{
+										Account:  "eosio.token",
+										Name:     "transfer",
+										JsonData: "",
+									},
+								},
+							},
+						},
+						PublicKeys: &PublicKeys{
+							PublicKeys: []string{"EOS5MHPYyhjBjnQZejzZHqHewPWhGTfQWSVTWYEhDmJu4SXkzgweP"},
+						},
+					}},
+				},
+				{
+					Id:           "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					BlockId:      "0000002ca98c204ec8f93e91332baf2a0eea5edd4e9f262bfb3e9b997d6f2415",
+					Irreversible: true,
+					Event: &TransactionEvent_Execution{Execution: &TransactionEvent_Executed{
+						BlockHeader: &BlockHeader{
+							ActionMroot: []byte("1"),
+						},
+						Trace: &TransactionTrace{
+							Receipt: &TransactionReceiptHeader{
+								Status: TransactionStatus_TRANSACTIONSTATUS_DELAYED,
+							},
+							Id:    "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+							Index: uint64(2),
+							RamOps: []*RAMOp{
+								{
+									Operation:   RAMOp_OPERATION_DEFERRED_TRX_PUSHED,
+									ActionIndex: 0,
+									Payer:       "eosio",
+									Delta:       371,
+									Usage:       1182644,
+									Namespace:   RAMOp_NAMESPACE_DEFERRED_TRX,
+									UniqueKey:   "9",
+									Action:      RAMOp_ACTION_PUSH,
+								},
+							},
+						},
+					}},
+				},
+				{
+					Id:           "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					BlockId:      "0000002ca98c204ec8f93e91332baf2a0eea5edd4e9f262bfb3e9b997d6f2415",
+					Irreversible: true,
+					Event: &TransactionEvent_DtrxScheduling{DtrxScheduling: &TransactionEvent_DtrxScheduled{
+						Transaction: &SignedTransaction{
+							Transaction: &Transaction{
+								Actions: []*Action{
+									{
+										Account:  "eosio.token",
+										Name:     "transfer",
+										JsonData: "{\"from\":\"eosio\",\"to\":\"battlefield1\",\"quantity\":\"1.0000 EOS\",\"memo\":\"push delayed trx\"}",
+									},
+								},
+							},
+						},
+						CreatedBy: &ExtDTrxOp{
+							SourceTransactionId: "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+							DtrxOp: &DTrxOp{
+								Operation: DTrxOp_OPERATION_PUSH_CREATE,
+							},
+						},
+					}},
+				},
+				{
+					Id:           "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					BlockId:      "0000002e7936c0363549d9f1bfc737e75a42a02d9e6cf72437e347f36580cfd8",
+					Irreversible: true,
+					Event: &TransactionEvent_Execution{Execution: &TransactionEvent_Executed{
+						BlockHeader: &BlockHeader{
+							ActionMroot: []byte("2"),
+						},
+						Trace: &TransactionTrace{
+							Receipt: &TransactionReceiptHeader{
+								Status: TransactionStatus_TRANSACTIONSTATUS_EXECUTED,
+							},
+							Id:    "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+							Index: uint64(1),
+							RamOps: []*RAMOp{
+								{
+									Operation:   RAMOp_OPERATION_DEFERRED_TRX_REMOVED,
+									ActionIndex: 0,
+									Payer:       "eosio",
+									Delta:       -371,
+									Usage:       1182273,
+									Namespace:   RAMOp_NAMESPACE_DEFERRED_TRX,
+									UniqueKey:   "9",
+									Action:      RAMOp_ACTION_REMOVE,
+								},
+							},
+						},
+					}},
+				},
+			},
+			canonicalChain: func(t *testing.T, id string) bool {
+				return true
+			},
+			expect: &TransactionLifecycle{
+				Id:                 "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+				TransactionStatus:  TransactionStatus_TRANSACTIONSTATUS_EXECUTED,
+				TransactionReceipt: &TransactionReceipt{Index: uint64(1)},
+				PublicKeys:         []string{"EOS5MHPYyhjBjnQZejzZHqHewPWhGTfQWSVTWYEhDmJu4SXkzgweP"},
+				Transaction: &SignedTransaction{
+					Transaction: &Transaction{
+						Actions: []*Action{
+							{
+								Account:  "eosio.token",
+								Name:     "transfer",
+								JsonData: "{\"from\":\"eosio\",\"to\":\"battlefield1\",\"quantity\":\"1.0000 EOS\",\"memo\":\"push delayed trx\"}",
+							},
+						},
+					},
+				},
+				CreatedBy: &ExtDTrxOp{
+					SourceTransactionId: "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					DtrxOp: &DTrxOp{
+						Operation: DTrxOp_OPERATION_PUSH_CREATE,
+					},
+				},
+				ExecutionBlockHeader: &BlockHeader{
+					ActionMroot: []byte("2"),
+				},
+				ExecutionTrace: &TransactionTrace{
+					Receipt: &TransactionReceiptHeader{
+						Status: TransactionStatus_TRANSACTIONSTATUS_EXECUTED,
+					},
+					Id:    "480a4adde14100097abec586d1dec805b3bfdb48c9efed5695ca02c61ea043bd",
+					Index: 1,
+					RamOps: []*RAMOp{
+						{
+							Operation:   RAMOp_OPERATION_DEFERRED_TRX_PUSHED,
+							ActionIndex: 0,
+							Payer:       "eosio",
+							Delta:       371,
+							Usage:       1182644,
+							Namespace:   RAMOp_NAMESPACE_DEFERRED_TRX,
+							UniqueKey:   "9",
+							Action:      RAMOp_ACTION_PUSH,
+						},
+						{
+							Operation:   RAMOp_OPERATION_DEFERRED_TRX_REMOVED,
+							ActionIndex: 0,
+							Payer:       "eosio",
+							Delta:       -371,
+							Usage:       1182273,
+							Namespace:   RAMOp_NAMESPACE_DEFERRED_TRX,
+							UniqueKey:   "9",
+							Action:      RAMOp_ACTION_REMOVE,
+						},
+					},
+				},
+				CreationIrreversible:  true,
 				ExecutionIrreversible: true,
 			},
 		},
@@ -154,4 +335,34 @@ func TestMergeTransactionEvents(t *testing.T) {
 			assert.Equal(t, test.expect, res)
 		})
 	}
+}
+
+func Test_deepMergeTransactionTrace(t *testing.T) {
+	a := &TransactionTrace{
+		Id:        "trx1",
+		RamOps:    []*RAMOp{{Operation: RAMOp_OPERATION_DEFERRED_TRX_REMOVED}},
+		RlimitOps: []*RlimitOp{{Operation: RlimitOp_OPERATION_INSERT}},
+	}
+	b := &TransactionTrace{
+		Id:        "trx1",
+		RamOps:    []*RAMOp{{Operation: RAMOp_OPERATION_DEFERRED_TRX_PUSHED}},
+		RlimitOps: []*RlimitOp{{Operation: RlimitOp_OPERATION_UPDATE}},
+	}
+	r1 := mergeTransactionTrace(a, b)
+	assert.Equal(t, []*RAMOp{{Operation: RAMOp_OPERATION_DEFERRED_TRX_PUSHED}, {Operation: RAMOp_OPERATION_DEFERRED_TRX_REMOVED}}, r1.RamOps)
+	assert.Equal(t, []*RlimitOp{{Operation: RlimitOp_OPERATION_UPDATE}, {Operation: RlimitOp_OPERATION_INSERT}}, r1.RlimitOps)
+
+	c := &TransactionTrace{
+		Id:        "trx1",
+		RamOps:    []*RAMOp{{Operation: RAMOp_OPERATION_DEFERRED_TRX_REMOVED}},
+		RlimitOps: []*RlimitOp{{Operation: RlimitOp_OPERATION_INSERT}},
+	}
+
+	d := &TransactionTrace{
+		Id: "trx1",
+	}
+	r2 := mergeTransactionTrace(c, d)
+	assert.Equal(t, []*RAMOp{{Operation: RAMOp_OPERATION_DEFERRED_TRX_REMOVED}}, r2.RamOps)
+	assert.Equal(t, []*RlimitOp{{Operation: RlimitOp_OPERATION_INSERT}}, r2.RlimitOps)
+
 }